student api bugfixes

+++ back-end/src/controller/student-controller.ts
@@ -7,27 +7,56 @@ export class StudentController {
   private studentRepository = getRepository(Student)
 
   async allStudents(request: Request, response: Response, next: NextFunction) {
-    return this.studentRepository.find()
+    try {
+      return this.studentRepository.find()
+    } catch (err) {
+      next(err)
+    }
+  }
+
+  async getStudent(request: Request, response: Response, next: NextFunction) {
+    try {
+      const { params } = request
+
+      const student = await this.studentRepository.findOne({
+        where: {
+          id: params.id,
+        },
+      })
+
+      if (!student) {
+        throw new Error(`student with id ${params.id} not found`)
+      }
+
+      return student
+    } catch (err) {
+      next(err)
+    }
   }
 
   async createStudent(request: Request, response: Response, next: NextFunction) {
-    const { body: params } = request
+    try {
+      const { body: params } = request
 
-    const createStudentInput: CreateStudentInput = {
-      first_name: params.first_name,
-      last_name: params.last_name,
-      photo_url: params.photo_url,
+      const createStudentInput: CreateStudentInput = {
+        first_name: params.first_name,
+        last_name: params.last_name,
+        photo_url: params.photo_url,
+      }
+      const student = new Student()
+      student.prepareToCreate(createStudentInput)
+
+      return this.studentRepository.save(student)
+    } catch (err) {
+      next(err)
     }
-    const student = new Student()
-    student.prepareToCreate(createStudentInput)
-
-    return this.studentRepository.save(student)
   }
 
   async updateStudent(request: Request, response: Response, next: NextFunction) {
-    const { body: params } = request
+    try {
+      const { body: params } = request
 
-    this.studentRepository.findOne(params.id).then((student) => {
+      const student = await this.studentRepository.findOne(params.id)
       const updateStudentInput: UpdateStudentInput = {
         id: params.id,
         first_name: params.first_name,
@@ -37,11 +66,22 @@ export class StudentController {
       student.prepareToUpdate(updateStudentInput)
 
       return this.studentRepository.save(student)
-    })
+    } catch (err) {
+      next(err)
+    }
   }
 
   async removeStudent(request: Request, response: Response, next: NextFunction) {
-    let studentToRemove = await this.studentRepository.findOne(request.params.id)
-    await this.studentRepository.remove(studentToRemove)
+    try {
+      let studentToRemove = await this.studentRepository.findOne(request.params.id)
+
+      if (!studentToRemove) {
+        throw new Error(`student with id ${request.params.id} not found`)
+      }
+
+      return await this.studentRepository.remove(studentToRemove)
+    } catch (err) {
+      next(err)
+    }
   }
 }
